fix: - remaned files - fixed exports - removed specified query results and replaced with one default - changed graphql querier from graphql-some to native node-fetch
