+++ package.json
@@ -49,7 +49,6 @@
     "react:remove": "yarn remove --frozen-lockfile react --dev && yarn add --frozen-lockfile --peer react@^17.0.0"
   },
   "dependencies": {
-    "@apollo/client": "^3.3.6",
     "@ethersproject/abi": "^5.5.0",
     "@ethersproject/abstract-signer": "^5.5.0",
     "@ethersproject/bignumber": "^5.5.0",
@@ -62,8 +61,6 @@
     "@ethersproject/wallet": "^5.5.0",
     "@ethersproject/web": "^5.5.1",
     "ethers": "^5.5.1",
-    "graphql": "^16.5.0",
-    "graphql-request": "^4.3.0",
     "lodash": "^4.17.21",
     "lodash-es": "^4.17.21",
     "object-hash": "^3.0.0"
