+++ src/explorer/index.ts
@@ -1,3 +1,1 @@
-export { getBridgeTxnInfo } from "./send_query";
-export { buildTxnQuery } from "./gql/query_builder";
-export { BRIDGE_TXN_API_GRAPHQL_URL, QUERY_INFO } from "./gql/utils";
+export { getBridgeTxnInfo } from "./bridge_transactions/query";
